+++ Godeps/_workspace/src/github.com/cloudfoundry-incubator/candiedyaml/encode_test.go
@@ -16,10 +16,11 @@ package candiedyaml
 
 import (
 	"bytes"
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
 	"math"
 	"time"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
 )
 
 var _ = Describe("Encode", func() {
@@ -36,6 +37,22 @@ var _ = Describe("Encode", func() {
 			err := enc.Encode("abc")
 			Ω(err).ShouldNot(HaveOccurred())
 			Ω(buf.String()).Should(Equal(`abc
+`))
+		})
+
+		It("encodes strings with multilines", func() {
+			err := enc.Encode("a\nc")
+			Ω(err).ShouldNot(HaveOccurred())
+			Ω(buf.String()).Should(Equal(`|-
+  a
+  c
+`))
+		})
+
+		It("handles strings that match known scalars", func() {
+			err := enc.Encode("true")
+			Ω(err).ShouldNot(HaveOccurred())
+			Ω(buf.String()).Should(Equal(`"true"
 `))
 		})