+++ Godeps/_workspace/src/github.com/cloudfoundry/loggregator_consumer/consumer_proxy_test.go
@@ -5,6 +5,7 @@ import (
 	"crypto/tls"
 	"errors"
 	consumer "github.com/cloudfoundry/loggregator_consumer"
+	"github.com/cloudfoundry/loggregatorlib/loggertesthelper"
 	"github.com/cloudfoundry/loggregatorlib/logmessage"
 	"github.com/cloudfoundry/loggregatorlib/server/handlers"
 	"github.com/elazarl/goproxy"
@@ -38,7 +39,7 @@ var _ = Describe("Loggregator Consumer behind a Proxy", func() {
 	BeforeEach(func() {
 		messagesToSend = make(chan []byte, 256)
 
-		testServer = httptest.NewServer(handlers.NewWebsocketHandler(messagesToSend, 100*time.Millisecond))
+		testServer = httptest.NewServer(handlers.NewWebsocketHandler(messagesToSend, 100*time.Millisecond, loggertesthelper.Logger()))
 		endpoint = "ws://" + testServer.Listener.Addr().String()
 		goProxyHandler = goproxy.NewProxyHttpServer()
 		goProxyHandler.Logger = log.New(bytes.NewBufferString(""), "", 0)
@@ -133,7 +134,7 @@ var _ = Describe("Loggregator Consumer behind a Proxy", func() {
 		}
 
 		BeforeEach(func() {
-			httpTestServer = httptest.NewServer(handlers.NewHttpHandler(messagesToSend))
+			httpTestServer = httptest.NewServer(handlers.NewHttpHandler(messagesToSend, loggertesthelper.Logger()))
 			endpoint = "ws://" + httpTestServer.Listener.Addr().String()
 		})