+++ Godeps/_workspace/src/github.com/cloudfoundry/noaa/errors/unauthorized_error.go
@@ -1,4 +1,4 @@
-package noaa_errors
+package errors
 
 type UnauthorizedError struct {
 	description string