+++ Godeps/_workspace/src/github.com/pivotal-cf-experimental/jibber_jabber/jibber_jabber_windows.go
@@ -21,8 +21,7 @@ var SUPPORTED_LOCALES = map[uintptr]string{
 	0x0416: "pt-BR",
 	//0x0419: "ru_RU", - Will add support for Russian when nicksnyder/go-i18n supports Russian
 	0x0804: "zh-CN",
-	0x0c04: "zh-HK",
-	0x0404: "zh-TW",
+	0x0c04: "zh-CK",
 }
 
 func getWindowsLocaleFrom(sysCall string) (locale string, err error) {